+++ src/Services/ThemeService/WidgetThemes/Dark.ts
@@ -1,0 +1,11 @@
+import { WidgetThemeDTO } from '@Types/ProjectTypes';
+
+export const DARK: WidgetThemeDTO = {
+  font_placeholder: '#999',
+  font:             '#EEE',
+  background:       '#444',
+  wrong:            '#F55',
+  confirm:          '#0C0',
+  modified:         '#AA0',
+  disabled:         '#888',
+};