+++ src/Services/ThemeService/WidgetThemes/index.ts
@@ -1,5 +1,7 @@
+import { DARK } from './Dark';
 import { DEFAULT } from './Default';
 
 export const widgetThemes = {
   DEFAULT: DEFAULT,
+  DARK: DARK,
 };