+++ docs/ReleaseNotes.md
@@ -1,3 +1,7 @@
 // Before any commit, use the command line: tsc --noEmit
 
-# 0.1.2.24 Added autofocus for modal textInput
+# 0.1.2.25 Widget Theme display
+
+- New Widget Theme Display implemented
+- Removed widget template buttons on Alert modal for widgets without name.
+- a minor reajust on pseudo widget children padding top.